+++ mbo/hash/hash_simple.h
@@ -13,20 +13,25 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-#ifndef MBO_TYPES_HASH_H_
-#define MBO_TYPES_HASH_H_
+#ifndef MBO_HASH_SIMPLE_HASH_H_
+#define MBO_HASH_SIMPLE_HASH_H_
+
+// IWYU pragma: private, include "mbo/hash/hash.h"
 
 #include <cstdint>
 #include <cstring>
 #include <string_view>
 #include <type_traits>
 
-namespace mbo::types {
+// Namespace `mbo::hash::simple` offers a `GetHash` a constexpr safe simple hash function.
+namespace mbo::hash::simple {
 
 // In theory this should be a random number or at least an arbitray number that
 // changes on every program start. Howeverm this number must be constexpr.
 static constexpr uint64_t kMaybeRandom = 5'008'709'998'333'326'415ULL;
 
+// A simple constexpr safe hash function. This function uses an optimized implementation at run-time
+// and a constexpr safe implementation that yields the exact same values for compile-time usage.
 inline constexpr uint64_t GetHash(std::string_view data) {
   // NOLINTBEGIN(*-magic-numbers)
   if (data.empty()) {
@@ -102,6 +107,6 @@ inline constexpr uint64_t GetHash(std::string_view data) {
   // NOLINTEND(*-magic-numbers)
 }
 
-}  // namespace mbo::types
+}  // namespace mbo::hash::simple
 
-#endif  // MBO_TYPES_HASH_H_
+#endif  // MBO_HASH_SIMPLE_HASH_H_