+++ mbo/types/tstring_test.cc
@@ -23,7 +23,7 @@
 #include "absl/hash/hash_testing.h"
 #include "gmock/gmock.h"
 #include "gtest/gtest.h"
-#include "mbo/types/hash.h"
+#include "mbo/hash/hash.h"
 
 namespace mbo::types {
 namespace {
@@ -658,13 +658,15 @@ TEST_F(TStringTest, FindFirstLast) {
 }
 
 TEST_F(TStringTest, StdHash) {
+  using ::mbo::hash::simple::GetHash;
   // "bar"_ts
   constexpr tstring<'b', 'a', 'r'> kTsBar;
   using TsBar = decltype(kTsBar);
   const std::size_t k_bar_hash_ts = std::hash<TsBar>{}(kTsBar);
   EXPECT_THAT(std::hash<TsBar>{}(), k_bar_hash_ts);
   constexpr auto kBarDirectHashTs = GetHash("bar");
-  EXPECT_THAT(kBarDirectHashTs, k_bar_hash_ts);
+  EXPECT_THAT(kBarDirectHashTs, Ne(k_bar_hash_ts));
+  EXPECT_THAT(kBarDirectHashTs, decltype(kTsBar)::StringHash());
   std::string_view vs_bar{"bar"};
   const auto k_bar_volatile_hash_ts = GetHash(vs_bar);
   EXPECT_THAT(k_bar_volatile_hash_ts, kBarDirectHashTs);
@@ -674,7 +676,8 @@ TEST_F(TStringTest, StdHash) {
   const std::size_t k_foo_hash_ts = std::hash<TsFoo>{}(kTsFoo);
   EXPECT_THAT(std::hash<TsFoo>{}(), k_foo_hash_ts);
   constexpr auto kFooDirectHashTs = GetHash("foo");
-  EXPECT_THAT(kFooDirectHashTs, k_foo_hash_ts);
+  EXPECT_THAT(kFooDirectHashTs, Ne(k_foo_hash_ts));
+  EXPECT_THAT(kFooDirectHashTs, decltype(kTsFoo)::StringHash());
   std::string_view vs_foo{"foo"};
   const auto k_foo_volatile_hash_ts = GetHash(vs_foo);
   EXPECT_THAT(k_foo_volatile_hash_ts, kFooDirectHashTs);