+++ CHANGELOG.md
@@ -1,6 +1,6 @@
 # 0.2.26
 
-* Added `GetHash(std::string_view)` which is constexpr safe.
+* Added `mbo::hash::simple::GetHash(std::string_view)` which is constexpr safe.
 * Added hash support to tstring.
 
 # 0.2.25